Move the Python tree lifecycle to platform/python_env.rs
ensure_user_python and everything it leans on: the version and defer
markers, the preserved version snapshot and restore, the symlink aware
copy, and interpreter_is_usable. mod.rs is left holding paths, PATH
manipulation, init and relaunch.
